+++ CHANGELOG.md
@@ -27,7 +27,6 @@ The main features of this release are:
     by replacing `sparse_poly.evaluate(pt)` to `sparse_poly.evaluate(&pt)`.
 - #129 (ark-ff) Move `ark_ff::{UniformRand, test_rng}` to `ark_std::{UniformRand, test_rng}`.
     Importing these from `ark-ff` is still possible, but is deprecated and will be removed in the following release.
-- #140 (ark-poly) Rename `Evaluations` to `UVEvaluations` 
 - #144 (ark-poly) Add `CanonicalSerialize` and `CanonicalDeserialize` trait bounds for `Polynomial`.
 - #160 (ark-serialize, ark-ff, ark-ec) 
   - Remove `ConstantSerializedSize`; users should use `serialized_size*` (see next).
@@ -46,6 +45,7 @@ The main features of this release are:
 - #96 (ark-ff) Make the `field_new` macro accept values in integer form, without requiring decomposition into limbs, and without requiring encoding in Montgomery form.
 - #106 (ark-ff, ark-ec) Add `Zeroize` trait bound to `Field, ProjectiveGroup, AffineGroup` traits.
 - #117 (ark-poly) Add operations to `SparsePolynomial`, so it implements `Polynomial`
+- #140 (ark-poly) Add support for multilinear extensions in dense and sparse evaluation form.
 
 ### Improvements
 - #22 (ark-ec) Speedup fixed-base MSMs
@@ -66,7 +66,6 @@ The main features of this release are:
 - #125 (ark-poly) Add parallelization to applying coset shifts within `coset_fft`.
 - #126 (ark-ec) Use `ark_ff::batch_inversion` for point normalization
 - #131, #137 (ark-ff) Speedup `sqrt` on fields when a square root exists. (And slows it down when doesn't exist)
-- #140 (ark-poly) Add support for multilinear extensions in dense and sparse evaluation form.
 - #141 (ark-ff) Add `Fp64`
 - #144 (ark-poly) Add serialization for polynomials and evaluations
 - #149 (ark-serialize) Add an impl of `CanonicalSerialize/Deserialize` for `String`.